calculate status func placed to Metrics
